+++ client/src/pages/Library/index.jsx
@@ -1,15 +1,21 @@
 import React from 'react'
 import './library.scss'
 import { connect } from 'react-redux'
-import BookItem from '../../components/BookItem'
 import { Grid } from 'semantic-ui-react'
+import BookItem from '../../components/BookItem'
 
-export const Library = ({ books }) => {
+//
+
+export const Library = ({ history, books }) => {
   return (
     <div className='library__container'>
       <Grid stackable container columns={1}>
         {books.map(book => (
-          <Grid.Column key={book.id} className='library__book'>
+          <Grid.Column
+            key={book.id}
+            className='library__book'
+            onClick={() => history.push(`/book/${book.id}`, { item: book })}
+          >
             <BookItem item={book} removeSave addRemove />
           </Grid.Column>
         ))}