+++ CHANGELOG.md
@@ -2,5 +2,4 @@
 
 ## CLIENT
 
-- show user books in the library
-- show only 300 chars from the book description
+- show individual books in the library